refactor(honor): the reply path moves to a file of its own
ActiveWalkViewModel+Honor.swift had reached the 500-line warning, and the
arrival reflection adds to exactly this group. No behaviour change:
ActiveWalkHonorTests and PilgrimageStageWalkTests pass untouched.
